backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
